move regression test to unit tests
